// src/api/models/Cover.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Cover
 */
export interface Cover {
    /**
     * 
     * @type {string}
     * @memberof Cover
     */
    cover_uri?: string;
    /**
     * 
     * @type {string}
     * @memberof Cover
     */
    cover_url?: string;
}

/**
 * Check if a given object implements the Cover interface.
 */
export function instanceOfCover(value: object): value is Cover {
    return true;
}

export function CoverFromJSON(json: any): Cover {
    return CoverFromJSONTyped(json, false);
}

export function CoverFromJSONTyped(json: any, ignoreDiscriminator: boolean): Cover {
    if (json == null) {
        return json;
    }
    return {
        
        'cover_uri': json['cover_uri'] == null ? undefined : json['cover_uri'],
        'cover_url': json['cover_url'] == null ? undefined : json['cover_url'],
    };
}

export function CoverToJSON(json: any): Cover {
    return CoverToJSONTyped(json, false);
}

export function CoverToJSONTyped(value?: Cover | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cover_uri': value['cover_uri'],
        'cover_url': value['cover_url'],
    };
}